Added additional comments

+++ src/PoorCodeCommenting.java
@@ -17,7 +17,7 @@ public class PoorCodeCommenting {
         loan = keyboard.nextDouble();
 
         System.out.printf("Enter the interest rate on the loan: ");
-        interest = keyboard.nextDouble();
+        interest = keyboard.nextInt();
 
         System.out.printf("Enter the term (years) for the loan payment: ");
         term = keyboard.nextInt();
@@ -25,14 +25,18 @@ public class PoorCodeCommenting {
         System.out.printf("\n======================================\n");
         keyboard.close();
 
-        //Formatting Output String for Balance Owed and Minimum Monthly Payments
+        //monthly_pay is set equal to the returned value in processSomething
         monthly_pay = processSomething(loan, interest, term);
+
+        //balance is set equal to monthly_pay times the term
         balance = -(monthly_pay*(term*12));
+
+        //Formatting Output String and double variable for Balance Owed and Minimum Monthly Payments
         System.out.format("%-30s$%-+10.2f%n", "Balance owed to bank: ", balance);
         System.out.format("%-30s$%-+10.2f%n", "Minimum monthly payment: ", monthly_pay);
     }
 
-
+    //A new class that returns the result value so it can be used elsewhere
     public static double processSomething(double loan, double interest, int term) {
         // calculates the monthly payment 
         double rate = (interest/ 100) / 12;
